Create initial Ship and Player classes
Added inital attributes and methods have been created for both the Ship
and Player class, as required from the project specification.

Ship has 'position' and 'sunk' attributes to easily describe the object
and its current status. While Player has a list of Ship objects that it
can manage with its own methods.

Created get_ship_coordinates(), guess_position(), is_sunk(), and
set_ship_coordinates() methods to Ship. The former three returns the
values of Ship for outside uses, such as the Player class. And
set_coordinates() changes the ship coordinate attribute to a desired
coordinate.

Made get_ship(), guess_position(), still_has_ship(), and add_ships()
method to Player. Again, the former three is used to get the status of
the ship objects inside of Player. add_ships() creates a new ship
object and places it inside of Player, only if it doesn't overlap with
another ship that has the same coordinates already in Player, as to not
create duplicate ships that already exist.